Updated PetStats and PetType to use quint8/16 when possible. PetAction revamped, consists of all moves that can be made. Move now uses PetAction and stores the move that provide that heuristic. All classes no longer use qDeleteAll and use another system of removing pointer objects from a QList. Added Polymorphed and Swapping to PetStatus. Added temp function to AI class to work as Expectiminimax while bugs are worked out. Set up new function CalculatePriority in AI class to make the main function look less bulky, will be used to determine who starts the round.

+++ WoWPetBattler/PetStats.cpp
@@ -1,25 +1,25 @@
 #include "PetStats.h"
 
 //Calculate the pet's health.
-int PetStats::CalculateHealth(float baseHealth, float breedHealth, float level, float quality)
+quint16 PetStats::CalculateHealth(float baseHealth, float breedHealth, float level, float quality)
 {
 	return Round((baseHealth + breedHealth) * 5 * level * (1 + ((quality - 1) / 10)) + 100);
 }
 
 //Calculate the pet's power.
-int PetStats::CalculatePower(float basePower, float breedPower, float level, float quality)
+quint16 PetStats::CalculatePower(float basePower, float breedPower, float level, float quality)
 {
 	return Round((basePower + breedPower) * level * (1 + ((quality - 1) / 10)));
 }
 
 //Calculate the pet's speed.
-int PetStats::CalculateSpeed(float baseSpeed, float breedSpeed, float level, float quality)
+quint16 PetStats::CalculateSpeed(float baseSpeed, float breedSpeed, float level, float quality)
 {
 	return Round((baseSpeed + breedSpeed) * level * (1 + ((quality - 1) / 10)));
 }
 
 //Rounds up or down depending on the decimal.
-int PetStats::Round(float value)
+quint16 PetStats::Round(float value)
 {
 	return qFloor(value + 0.5);
 }
