+++ WoWPetBattler/PetStats.h
@@ -1,17 +1,18 @@
 #ifndef PETSTATS_H
 #define PETSTATS_H
 
+#include <QDataStream>
 #include <QtCore/qmath.h>
 
 class PetStats
 {
 public:
-	static int CalculateHealth(float, float, float, float);
-	static int CalculatePower(float, float, float, float);
-	static int CalculateSpeed(float, float, float, float);
+	static quint16 CalculateHealth(float, float, float, float);
+	static quint16 CalculatePower(float, float, float, float);
+	static quint16 CalculateSpeed(float, float, float, float);
 
 private:
-	static int Round(float);
+	static quint16 Round(float);
 };
 
 #endif
